+++ src/simpleGit.ts
@@ -220,9 +220,9 @@ export class SimpleGit extends GitManager {
         this.plugin.setState(PluginState.idle);
     }
 
-    async unstageAll(): Promise<void> {
+    async unstageAll({ dir }: { dir?: string; }): Promise<void> {
         this.plugin.setState(PluginState.add);
-        await this.git.reset([], (err) => this.onError(err));
+        await this.git.reset(dir != undefined ? ["--", dir] : [], (err) => this.onError(err));
         this.plugin.setState(PluginState.idle);
     }
 
@@ -260,6 +260,10 @@ export class SimpleGit extends GitManager {
         return revision;
     }
 
+    async discardAll({ dir }: { dir?: string; }): Promise<void> {
+        return this.discard(dir ?? ".");
+    }
+
     async pull(): Promise<FileStatusResult[] | undefined> {
         this.plugin.setState(PluginState.pull);
         if (this.plugin.settings.updateSubmodules)
@@ -290,7 +294,7 @@ export class SimpleGit extends GitManager {
             } else if (this.plugin.settings.syncMethod === 'reset') {
                 try {
                     await this.git.raw(['update-ref', `refs/heads/${branchInfo.current}`, upstreamCommit], (err) => this.onError(err));
-                    await this.unstageAll();
+                    await this.unstageAll({});
                 } catch (err) {
                     this.plugin.displayError(`Sync failed (${this.plugin.settings.syncMethod}): ${err.message}`);
                 }