+++ src/ui/sidebar/gitView.svelte
@@ -10,6 +10,7 @@
 	} from "src/types";
 	import { onDestroy } from "svelte";
 	import { slide } from "svelte/transition";
+	import { DiscardModal } from "../modals/discardModal";
 	import FileComponent from "./components/fileComponent.svelte";
 	import PulledFileComponent from "./components/pulledFileComponent.svelte";
 	import StagedFileComponent from "./components/stagedFileComponent.svelte";
@@ -87,6 +88,7 @@
 			lastPulledFilesHierarchy = {
 				title: "",
 				path: "",
+				vaultPath: "",
 				children: plugin.gitManager.getTreeStructure(lastPulledFiles),
 			};
 		}
@@ -108,6 +110,7 @@
 				changeHierarchy = {
 					title: "",
 					path: "",
+					vaultPath: "",
 					children: plugin.gitManager.getTreeStructure(
 						status.changed
 					),
@@ -115,6 +118,7 @@
 				stagedHierarchy = {
 					title: "",
 					path: "",
+					vaultPath: "",
 					children: plugin.gitManager.getTreeStructure(status.staged),
 				};
 			}
@@ -149,6 +153,23 @@
 		loading = true;
 		plugin.pullChangesFromRemote().finally(triggerRefresh);
 	}
+	function discard() {
+		new DiscardModal(view.app, false, plugin.gitManager.getVaultPath("/"))
+			.myOpen()
+			.then((shouldDiscard) => {
+				if (shouldDiscard === true) {
+					plugin.gitManager
+						.discardAll({
+							status: plugin.cachedStatus,
+						})
+						.finally(() => {
+							dispatchEvent(new CustomEvent("git-refresh"));
+						});
+				}
+			});
+	}
+
+	$: rows = (commitMessage.match(/\n/g) || []).length + 1 || 1;
 </script>
 
 <main>
@@ -211,6 +232,7 @@
 				class:loading
 				data-icon="refresh-cw"
 				aria-label="Refresh"
+				style="margin: 1px;"
 				bind:this={buttons[6]}
 				on:click={triggerRefresh}
 			/>
@@ -218,6 +240,7 @@
 	</div>
 	<div class="git-commit-msg">
 		<textarea
+			{rows}
 			class="commit-msg-input"
 			type="text"
 			spellcheck="true"
@@ -243,10 +266,11 @@
 					>
 						<div
 							class="nav-folder-title"
-							on:click={() => (stagedOpen = !stagedOpen)}
+							on:click|self={() => (stagedOpen = !stagedOpen)}
 						>
 							<div
 								class="nav-folder-collapse-indicator collapse-icon"
+								on:click={() => (stagedOpen = !stagedOpen)}
 							>
 								<svg
 									xmlns="http://www.w3.org/2000/svg"
@@ -262,12 +286,45 @@
 									><path d="M3 8L12 17L21 8" /></svg
 								>
 							</div>
-							<div class="nav-folder-title-content">
+							<div
+								class="nav-folder-title-content"
+								on:click={() => (stagedOpen = !stagedOpen)}
+							>
 								Staged Changes
 							</div>
-							<span class="tree-item-flair"
-								>{status.staged.length}</span
-							>
+
+							<div class="tools">
+								<div class="buttons">
+									<div
+										data-icon="minus"
+										aria-label="Unstage"
+										bind:this={buttons[8]}
+										on:click={unstageAll}
+										class="clickable-icon"
+									>
+										<svg
+											width="18"
+											height="18"
+											viewBox="0 0 18 18"
+											fill="none"
+											stroke="currentColor"
+											stroke-width="2"
+											stroke-linecap="round"
+											stroke-linejoin="round"
+											class="svg-icon lucide-minus"
+											><line
+												x1="4"
+												y1="9"
+												x2="14"
+												y2="9"
+											/></svg
+										>
+									</div>
+								</div>
+								<div class="files-count">
+									{status.staged.length}
+								</div>
+							</div>
 						</div>
 						{#if stagedOpen}
 							<div
@@ -299,11 +356,12 @@
 						class:is-collapsed={!changesOpen}
 					>
 						<div
-							on:click={() => (changesOpen = !changesOpen)}
+							on:click|self={() => (changesOpen = !changesOpen)}
 							class="nav-folder-title"
 						>
 							<div
 								class="nav-folder-collapse-indicator collapse-icon"
+								on:click={() => (changesOpen = !changesOpen)}
 							>
 								<svg
 									xmlns="http://www.w3.org/2000/svg"
@@ -320,10 +378,72 @@
 								>
 							</div>
 
-							<div class="nav-folder-title-content">Changes</div>
-							<span class="tree-item-flair"
-								>{status.changed.length}</span
+							<div
+								class="nav-folder-title-content"
+								on:click={() => (changesOpen = !changesOpen)}
 							>
+								Changes
+							</div>
+							<div class="tools">
+								<div class="buttons">
+									<div
+										data-icon="undo"
+										aria-label="Discard"
+										on:click={() => discard()}
+										class="clickable-icon"
+									>
+										<svg
+											xmlns="http://www.w3.org/2000/svg"
+											width="24"
+											height="24"
+											viewBox="0 0 24 24"
+											fill="none"
+											stroke="currentColor"
+											stroke-width="2"
+											stroke-linecap="round"
+											stroke-linejoin="round"
+											class="svg-icon lucide-undo"
+											><path d="M3 7v6h6" /><path
+												d="M21 17a9 9 0 0 0-9-9 9 9 0 0 0-6 2.3L3 13"
+											/></svg
+										>
+									</div>
+									<div
+										data-icon="plus"
+										aria-label="Stage"
+										bind:this={buttons[9]}
+										on:click={stageAll}
+										class="clickable-icon"
+									>
+										<svg
+											xmlns="http://www.w3.org/2000/svg"
+											width="24"
+											height="24"
+											viewBox="0 0 24 24"
+											fill="none"
+											stroke="currentColor"
+											stroke-width="2"
+											stroke-linecap="round"
+											stroke-linejoin="round"
+											class="svg-icon lucide-plus"
+											><line
+												x1="12"
+												y1="5"
+												x2="12"
+												y2="19"
+											/><line
+												x1="5"
+												y1="12"
+												x2="19"
+												y2="12"
+											/></svg
+										>
+									</div>
+								</div>
+								<div class="files-count">
+									{status.changed.length}
+								</div>
+							</div>
 						</div>
 						{#if changesOpen}
 							<div
@@ -423,9 +543,8 @@
 <style lang="scss">
 	.commit-msg-input {
 		width: 100%;
-		min-height: 33px;
-		height: 30px;
-		resize: vertical;
+		overflow: hidden;
+		resize: none;
 		padding: 7px 5px;
 		background-color: var(--background-modifier-form-field);
 	}
@@ -436,6 +555,26 @@
 		width: calc(100% - var(--size-4-8));
 		margin: 4px auto;
 	}
+	main {
+		.tools {
+			display: flex;
+			margin-left: auto;
+			.buttons {
+				display: flex;
+				> * {
+					padding: 0 0;
+					height: auto;
+				}
+			}
+			.files-count {
+				padding-left: var(--size-2-1);
+				width: 11px;
+				display: flex;
+				align-items: center;
+				justify-content: center;
+			}
+		}
+	}
 
 	.git-commit-msg-clear-button {
 		position: absolute;
@@ -443,7 +582,7 @@
 		border-radius: 50%;
 		color: var(--search-clear-button-color);
 		cursor: var(--cursor);
-		top: 0px;
+		top: -4px;
 		right: 2px;
 		bottom: 0px;
 		line-height: 0;